# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wisma Eucaliptus / Guest House Deluxe - Property Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with 4-digit PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify:
    get:
      operationId: verifySession
      tags: [auth]
      summary: Verify session token
      responses:
        "200":
          description: Session valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid or expired session
...
  /guests:
    get:
      operationId: getGuests
      tags: [guests]
      summary: List guests (optionally filter active/all)
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Create a guest record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Guest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get a single guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guest detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "404":
          description: Guest not found
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update guest info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: List bookings (filter by status, date range)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [reserved, checked_in, checked_out, cancelled]
        - name: room_id
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking / advance reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a single booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Cancel/delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /bookings/{id}/checkin:
    post:
      operationId: checkIn
      tags: [bookings]
      summary: Check in a guest (changes status to checked_in)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/checkout:
    post:
      operationId: checkOut
      tags: [bookings]
      summary: Check out a guest (changes status to checked_out)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutInput"
      responses:
        "200":
          description: Checked out with receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckOutReceipt"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginInput:
      type: object
      properties:
        pin:
          type: string
          minLength: 4
          maxLength: 4
      required:
        - pin
    AuthResult:
      type: object
      properties:
        token:
          type: string
        staff:
          type: string
      required:
        - token
        - staff
...
    Guest:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        id_number:
          type: string
        id_type:
          type: string
          enum: [ktp, passport]
        nationality:
          type: string
        phone:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - id_number
        - id_type
        - nationality
        - created_at
    GuestInput:
      type: object
      properties:
        name:
          type: string
        id_number:
          type: string
        id_type:
          type: string
          enum: [ktp, passport]
        nationality:
          type: string
        phone:
          type: string
      required:
        - name
        - id_number
        - id_type
        - nationality
    GuestUpdate:
      type: object
      properties:
        name:
          type: string
        id_number:
          type: string
        id_type:
          type: string
        nationality:
          type: string
        phone:
          type: ["string", "null"]
    Booking:
      type: object
      properties:
        id:
          type: integer
        room_id:
          type: integer
        room_number:
          type: string
        guest_id:
          type: integer
        guest_name:
          type: string
        guest_nationality:
          type: string
        check_in_date:
          type: string
          format: date
        check_out_date:
          type: string
          format: date
        actual_check_in:
          type: ["string", "null"]
          format: date-time
        actual_check_out:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [reserved, checked_in, checked_out, cancelled]
        stay_type:
          type: string
          enum: [regular, long_stay]
        price_per_night:
          type: integer
        total_amount:
          type: ["integer", "null"]
        deposit:
          type: ["integer", "null"]
        payment_method:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - room_id
        - room_number
        - guest_id
        - guest_name
        - guest_nationality
        - check_in_date
        - check_out_date
        - status
        - stay_type
        - price_per_night
        - created_at
    BookingInput:
      type: object
      properties:
        room_id:
          type: integer
        guest_id:
          type: ["integer", "null"]
        guest:
          $ref: "#/components/schemas/GuestInput"
        check_in_date:
          type: string
          format: date
        check_out_date:
          type: string
          format: date
        stay_type:
          type: string
          enum: [regular, long_stay]
        price_per_night:
          type: integer
        deposit:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - room_id
        - check_in_date
        - check_out_date
        - stay_type
        - price_per_night
    BookingUpdate:
      type: object
      properties:
        check_in_date:
          type: string
          format: date
        check_out_date:
          type: string
          format: date
        price_per_night:
          type: integer
        deposit:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [reserved, checked_in, checked_out, cancelled]
    CheckInInput:
      type: object
      properties:
        deposit:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    CheckOutInput:
      type: object
      properties:
        payment_method:
          type: string
          enum: [cash, transfer, card]
        additional_charges:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - payment_method
    CheckOutReceipt:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        nights:
          type: integer
        subtotal:
          type: integer
        additional_charges:
          type: integer
        deposit_paid:
          type: integer
        total_due:
          type: integer
        payment_method:
          type: string
      required:
        - booking
        - nights
        - subtotal
        - additional_charges
        - deposit_paid
        - total_due
        - payment_method
...