# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wisma Eucaliptus / Guest House Deluxe - Property Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthResult {
  token: string;
  staff: string;
}