# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wisma Eucaliptus / Guest House Deluxe - Property Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';
import type { BookingStayType } from './bookingStayType';

export interface Booking {
  id: number;
  room_id: number;
  room_number: string;
  guest_id: number;
  guest_name: string;
  guest_nationality: string;
  check_in_date: Date;
  check_out_date: Date;
  /** @nullable */
  actual_check_in?: Date | null;
  /** @nullable */
  actual_check_out?: Date | null;
  status: BookingStatus;
  stay_type: BookingStayType;
  price_per_night: number;
  /** @nullable */
  total_amount?: number | null;
  /** @nullable */
  deposit?: number | null;
  /** @nullable */
  payment_method?: string | null;
  /** @nullable */
  notes?: string | null;
  created_at: Date;
}