# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wisma Eucaliptus / Guest House Deluxe - Property Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputStayType } from './bookingInputStayType';
import type { GuestInput } from './guestInput';

export interface BookingInput {
  room_id: number;
  /** @nullable */
  guest_id?: number | null;
  guest?: GuestInput;
  check_in_date: Date;
  check_out_date: Date;
  stay_type: BookingInputStayType;
  price_per_night: number;
  /** @nullable */
  deposit?: number | null;
  /** @nullable */
  notes?: string | null;
}