# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wisma Eucaliptus / Guest House Deluxe - Property Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingUpdateStatus } from './bookingUpdateStatus';

export interface BookingUpdate {
  check_in_date?: Date;
  check_out_date?: Date;
  price_per_night?: number;
  /** @nullable */
  deposit?: number | null;
  /** @nullable */
  notes?: string | null;
  status?: BookingUpdateStatus;
}