# lib/api-zod/src/generated/types/checkOutInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wisma Eucaliptus / Guest House Deluxe - Property Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckOutInputPaymentMethod } from './checkOutInputPaymentMethod';

export interface CheckOutInput {
  payment_method: CheckOutInputPaymentMethod;
  /** @nullable */
  additional_charges?: number | null;
  /** @nullable */
  notes?: string | null;
}