# lib/api-zod/src/generated/types/guestUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wisma Eucaliptus / Guest House Deluxe - Property Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface GuestUpdate {
  name?: string;
  id_number?: string;
  id_type?: string;
  nationality?: string;
  /** @nullable */
  phone?: string | null;
}